chore: ensure EOF newline in index.html
