+++ vite.config.ts
@@ -4,10 +4,14 @@ import { fileURLToPath, URL } from 'node:url';
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  base: '/HeavyOrc/',
   plugins: [react()],
   resolve: {
     alias: {
       '@': fileURLToPath(new URL('./', import.meta.url)),
     },
+  },
+  build: {
+    target: 'es2022',
   },
 });
